+++ sdk/core/azure-core/src/http/curl/curl.cpp
@@ -186,61 +186,69 @@ static void CleanupThread()
   using namespace Azure::Core::Http::_detail;
   for (;;)
   {
+    Log::Write(Logger::Level::Verbose, "Clean pool check now...");
+    // Won't continue until the ConnectionPoolMutex is released from MoveConnectionBackToPool
+    std::unique_lock<std::mutex> lockForPoolCleaning(
+        CurlConnectionPool::g_curlConnectionPool.ConnectionPoolMutex);
+    Log::Write(Logger::Level::Verbose, "Clean pool sleep");
+    // Wait for the default time OR to the signal from the conditional variable.
+    // wait_for releases the mutex lock when it goes to sleep and it takes the lock again when it
+    // wakes up (or it's cancelled).
+    if (CurlConnectionPool::g_curlConnectionPool.ConditionalVariableForCleanThread.wait_for(
+            lockForPoolCleaning,
+            std::chrono::milliseconds(DefaultCleanerIntervalMilliseconds),
+            []() {
+              return CurlConnectionPool::g_curlConnectionPool.ConnectionPoolIndex.size() == 0;
+            }))
     {
-      // Won't continue until the ConnectionPoolMutex is released from MoveConnectionBackToPool
-      std::unique_lock<std::mutex> lockForPoolCleaning(
-          CurlConnectionPool::g_curlConnectionPool.ConnectionPoolMutex);
-      // Wait the defined default time OR to the signal from the conditional variable.
-      // wait_until releases the mutex lock until it wakes up again or it's cancelled.
-      if (CurlConnectionPool::g_curlConnectionPool.ConditionalVariableForCleanThread.wait_for(
-              lockForPoolCleaning,
-              std::chrono::milliseconds(DefaultCleanerIntervalMilliseconds),
-              []() { return CurlConnectionPool::g_curlConnectionPool.ConnectionCounter == 0; }))
+      // Cancelled by another thead or no connections on wakeup
+      Log::Write(
+          Logger::Level::Verbose,
+          "Clean pool - no connections on wake - return *************************");
+      CurlConnectionPool::g_curlConnectionPool.IsCleanThreadRunning = false;
+      return;
+    }
+
+    Log::Write(Logger::Level::Verbose, "Clean pool - inspect pool");
+    // loop the connection pool index - Note: lock is re-taken for the mutex
+    // Notes: The size of each host-index is always expected to be greater than 0 because the
+    // host-index is removed anytime it becomes empty.
+    for (auto index = CurlConnectionPool::g_curlConnectionPool.ConnectionPoolIndex.begin();
+         index != CurlConnectionPool::g_curlConnectionPool.ConnectionPoolIndex.end();)
+    {
+      // Each pool index behaves as a Last-in-First-out (connections are added to the pool with
+      // push_front). The last connection moved to the pool will be the first to be re-used. Because
+      // of this, the oldest connection in the pool can be found at the end of the list. Looping the
+      // connection pool backwards until a connection that is not expired is found or until all
+      // connections are removed.
+      for (auto connection = --(index->second.end());
+           index->second.size() > 0 && connection->get()->IsExpired();
+           connection = index->second.size() > 0 ? --connection : connection)
       {
-        // Cancelled by another thead or no connections on wakeup
-        return;
+        // remove connection from the pool and update the connection to the next one
+        // which is going to be list.end()
+        Log::Write(Logger::Level::Verbose, "Clean pool - remove connection");
+        connection = index->second.erase(connection);
       }
 
-      // loop the connection pool index - Note: lock is re-taken for the mutex
-      for (auto index = CurlConnectionPool::g_curlConnectionPool.ConnectionPoolIndex.begin();
-           index != CurlConnectionPool::g_curlConnectionPool.ConnectionPoolIndex.end();
-           index++)
+      if (index->second.size() == 0)
       {
-        if (index->second.size() == 0)
-        {
-          // Move the next pool index
-          continue;
-        }
-
-        // Pool index with waiting connections. Loop the connection pool backwards until
-        // a connection that is not expired is found or until all connections are removed.
-        for (auto connection = index->second.end();;)
-        {
-          // loop starts at end(), go back to previous possition. We know the list is
-          // size() > 0 so we are safe to go end() - 1 and find the last element in the
-          // list
-          connection--;
-          if (connection->get()->IsExpired())
-          {
-            // remove connection from the pool and update the connection to the next one
-            // which is going to be list.end()
-            connection = index->second.erase(connection);
-            CurlConnectionPool::g_curlConnectionPool.ConnectionCounter -= 1;
-
-            // Connection removed, break if there are no more connections to check
-            if (index->second.size() == 0)
-            {
-              break;
-            }
-          }
-          else
-          {
-            // Got a non-expired connection, all connections before this one are not
-            // expired. Break the loop and continue looping the Pool index
-            break;
-          }
-        }
+        Log::Write(Logger::Level::Verbose, "Clean pool - remove index " + index->first);
+        index = CurlConnectionPool::g_curlConnectionPool.ConnectionPoolIndex.erase(index);
       }
+      else
+      {
+        index = ++index;
+      }
+    }
+
+    if (CurlConnectionPool::g_curlConnectionPool.ConnectionPoolIndex.size() == 0)
+    {
+      Log::Write(
+          Logger::Level::Verbose,
+          "Clean pool - all connections removed. Return**********************");
+      CurlConnectionPool::g_curlConnectionPool.IsCleanThreadRunning = false;
+      return;
     }
   }
 }
@@ -1218,7 +1226,9 @@ std::unique_ptr<CurlNetworkConnection> CurlConnectionPool::ExtractOrCreateCurlCo
     CurlTransportOptions const& options,
     bool resetPool)
 {
-  std::string const& host = request.GetUrl().GetHost();
+  uint16_t port = request.GetUrl().GetPort();
+  std::string const& host = request.GetUrl().GetScheme() + request.GetUrl().GetHost()
+      + (port != 0 ? std::to_string(port) : "");
   std::string const connectionKey = GetConnectionKey(host, options);
 
   {
@@ -1234,8 +1244,9 @@ std::unique_ptr<CurlNetworkConnection> CurlConnectionPool::ExtractOrCreateCurlCo
     {
       if (resetPool)
       {
-        // Remove all connections for the connection Key and move to spawn new connection below
-        g_curlConnectionPool.ConnectionCounter -= hostPoolIndex->second.size();
+        // clean the pool-index as requested in the call. Typically to force a new connection to be
+        // created and to discard all current connections in the pool for the host-index. A caller
+        // might request this after getting broken/closed connections multiple-times.
         hostPoolIndex->second.clear();
       }
       else
@@ -1246,8 +1257,6 @@ std::unique_ptr<CurlNetworkConnection> CurlConnectionPool::ExtractOrCreateCurlCo
         auto connection = std::move(*fistConnectionIterator);
         // Remove the connection ref from list
         hostPoolIndex->second.erase(fistConnectionIterator);
-        // reduce number of connections on the pool
-        g_curlConnectionPool.ConnectionCounter -= 1;
 
         // Remove index if there are no more connections
         if (hostPoolIndex->second.size() == 0)
@@ -1280,7 +1289,6 @@ std::unique_ptr<CurlNetworkConnection> CurlConnectionPool::ExtractOrCreateCurlCo
         + std::string(curl_easy_strerror(result)));
   }
 
-  uint16_t port = request.GetUrl().GetPort();
   if (port != 0 && !SetLibcurlOption(newHandle, CURLOPT_PORT, port, &result))
   {
     throw Azure::Core::Http::TransportException(
@@ -1384,19 +1392,41 @@ void CurlConnectionPool::MoveConnectionBackToPool(
     return;
   }
 
+  Log::Write(Logger::Level::Verbose, "Moving connection to pool...");
+
   // Lock mutex to access connection pool. mutex is unlock as soon as lock is out of scope
   std::lock_guard<std::mutex> lock(CurlConnectionPool::ConnectionPoolMutex);
   auto& poolId = connection->GetConnectionKey();
   auto& hostPool = g_curlConnectionPool.ConnectionPoolIndex[poolId];
+
+  if (hostPool.size() >= _detail::MaxConnectionsPerIndex)
+  {
+    // Remove the last connection from the pool to insert this one.
+    auto lastConnection = --hostPool.end();
+    hostPool.erase(lastConnection);
+  }
+
   // update the time when connection was moved back to pool
-  connection->updateLastUsageTime();
+  connection->UpdateLastUsageTime();
   hostPool.push_front(std::move(connection));
-  g_curlConnectionPool.ConnectionCounter += 1;
+
+  if (m_cleanThread.joinable() && !IsCleanThreadRunning)
+  {
+    // Clean thread was running before but it's finished, join it to finalize
+    m_cleanThread.join();
+  }
+
   // Cleanup will start a background thread which will close abandoned connections from the pool.
   // This will free-up resources from the app
   // This is the only call to cleanup.
   if (!m_cleanThread.joinable())
   {
+    Log::Write(Logger::Level::Verbose, "Start clean thread");
+    IsCleanThreadRunning = true;
     m_cleanThread = std::thread(CleanupThread);
+  }
+  else
+  {
+    Log::Write(Logger::Level::Verbose, "Clean thread running. Won't start a new one.");
   }
 }