+++ sdk/core/azure-core/src/http/curl/curl_connection_private.hpp
@@ -33,6 +33,9 @@ namespace Azure { namespace Core { namespace Http {
     constexpr static int DefaultCleanerIntervalMilliseconds = 1000 * 90;
     // 60 sec -> expired connection is when it waits for 60 sec or more and it's not re-used
     constexpr static int DefaultConnectionExpiredMilliseconds = 1000 * 60;
+    // Define the maximun allowed connections per host-index in the pool. If this number is reached
+    // for the host-index, next connections trying to be added to the pool will be ignored.
+    constexpr static size_t MaxConnectionsPerIndex = 1024;
   } // namespace _detail
 
   /**
@@ -62,7 +65,7 @@ namespace Azure { namespace Core { namespace Http {
     /**
      * @brief Update last usage time for the connection.
      */
-    virtual void updateLastUsageTime() = 0;
+    virtual void UpdateLastUsageTime() = 0;
 
     /**
      * @brief Checks whether this CURL connection is expired.
@@ -150,7 +153,7 @@ namespace Azure { namespace Core { namespace Http {
       /**
        * @brief Update last usage time for the connection.
        */
-      void updateLastUsageTime() override
+      void UpdateLastUsageTime() override
       {
         this->m_lastUseTime = std::chrono::steady_clock::now();
       }