+++ sdk/core/azure-core/test/ut/curl_options.cpp
@@ -90,8 +90,8 @@ namespace Azure { namespace Core { namespace Test {
 
     // Clean the connection from the pool *Windows fails to clean if we leave to be clean uppon
     // app-destruction
-    EXPECT_NO_THROW(
-        Azure::Core::Http::_detail::CurlConnectionPool::g_curlConnectionPool.ResetPool());
+    EXPECT_NO_THROW(Azure::Core::Http::_detail::CurlConnectionPool::g_curlConnectionPool
+                        .ConnectionPoolIndex.clear());
   }
 
   /*
@@ -209,8 +209,8 @@ namespace Azure { namespace Core { namespace Test {
 
     // Clean the connection from the pool *Windows fails to clean if we leave to be clean uppon
     // app-destruction
-    EXPECT_NO_THROW(
-        Azure::Core::Http::_detail::CurlConnectionPool::g_curlConnectionPool.ResetPool());
+    EXPECT_NO_THROW(Azure::Core::Http::_detail::CurlConnectionPool::g_curlConnectionPool
+                        .ConnectionPoolIndex.clear());
   }
 
   TEST(CurlTransportOptions, httpsDefault)
@@ -242,8 +242,8 @@ namespace Azure { namespace Core { namespace Test {
 
     // Clean the connection from the pool *Windows fails to clean if we leave to be clean upon
     // app-destruction
-    EXPECT_NO_THROW(
-        Azure::Core::Http::_detail::CurlConnectionPool::g_curlConnectionPool.ResetPool());
+    EXPECT_NO_THROW(Azure::Core::Http::_detail::CurlConnectionPool::g_curlConnectionPool
+                        .ConnectionPoolIndex.clear());
   }
 
   TEST(CurlTransportOptions, disableKeepAlive)
@@ -280,7 +280,9 @@ namespace Azure { namespace Core { namespace Test {
     }
     // Make sure there are no connections in the pool
     EXPECT_EQ(
-        Azure::Core::Http::_detail::CurlConnectionPool::g_curlConnectionPool.ConnectionCounter, 0);
+        Azure::Core::Http::_detail::CurlConnectionPool::g_curlConnectionPool.ConnectionPoolIndex
+            .size(),
+        0);
   }
 
 }}} // namespace Azure::Core::Test