+++ sdk/storage/azure-storage-files-shares/inc/azure/storage/files/shares/share_service_client.hpp
@@ -67,15 +67,14 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
     std::string GetUrl() const { return m_serviceUrl.GetAbsoluteUrl(); }
 
     /**
-     * @brief List the shares from the service.
+     * @brief Returns a paginated collection of the shares in the storage account. Enumerating the
+     * shares may make multiple requests to the service while fetching all the values.
      * @param options Optional parameters to list the shares.
      * @param context Context for cancelling long running operations.
-     * @return Azure::Response<Models::ListSharesSinglePageResult> The results containing the
-     * shares returned and information used for future list operation on valid result not yet
-     * returned.
+     * @return ListSharesPagedResponse describing the shares in this storage account.
      */
-    Azure::Response<Models::ListSharesSinglePageResult> ListSharesSinglePage(
-        const ListSharesSinglePageOptions& options = ListSharesSinglePageOptions(),
+    ListSharesPagedResponse ListShares(
+        const ListSharesOptions& options = ListSharesOptions(),
         const Azure::Core::Context& context = Azure::Core::Context()) const;
 
     /**