+++ sdk/storage/azure-storage-files-datalake/CHANGELOG.md
@@ -1,15 +1,23 @@
 # Release History
 
-## 12.0.0-beta.10 (Unreleased)
+## 12.0.0-beta.11 (Unreleased)
+
+
+## 12.0.0-beta.10 (2021-04-16)
 
 ### Breaking Changes
 
 - Removed `Azure::Storage::Files::DataLake::PackageVersion`.
-- Renamed `GetUserDelegationKeyOptions::startsOn` to `GetUserDelegationKeyOptions::StartsOn`.
+- Renamed `GetUserDelegationKeyOptions::startsOn` to `StartsOn`.
+- Replaced all paginated collection functions that have the SinglePage suffix with pageable functions returning a `PagedResponse<T>`-derived type. The options are also renamed accordingly.
+  - `DataLakeServiceClient::ListFileSystems()`.
+  - `DataLakeFileSystemClient::ListPaths()`.
+  - `DataLakeDirectoryClient::ListPaths()`.
+- Removed `DataLakePathClient::SetAccessControlListRecursiveSinglePage()`, `UpdateAccessControlListRecursiveSinglePage()` and `RemoveAccessControlListRecursiveSinglePage()`.
 
 ### Bug Fixes
 
-- Rename functions always fail because `/` was left out in rename source path.
+- Rename functions always fail because `/` was left out in the renamed source path.
 
 ## 12.0.0-beta.9 (2021-03-23)